chore(entities): more restructuring in test files
